Fixed a bug with BWAPI, Modified Strategymanager
I think I've fixed the problem we've had with upgrades blocking unit
production. Turns out there was no code to handle an addon that was
doing research. It still runs after my change, so we'll see.

Also fiddled with the build order some more. We now beat Ualbertabot's
zealot rush most of the time.
added a double bunker build order that I'm not sure if we need or not,
and isn't optimized at all.